requirements.txt
discord.py==1.5.1
dotenv==0.14.0
requests==2.25.0